+++ cmd/gcx/skills/command.go
@@ -5,24 +5,15 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/fs"
-	"os"
-	"path"
 	"path/filepath"
-	"sort"
 	"strings"
 	"text/tabwriter"
 
 	"github.com/grafana/gcx/internal/agent"
 	"github.com/grafana/gcx/internal/format"
 	cmdio "github.com/grafana/gcx/internal/output"
+	internalskills "github.com/grafana/gcx/internal/skills"
 	"github.com/spf13/cobra"
-	"gopkg.in/yaml.v3"
-)
-
-const (
-	skillsSourceEnv = "GCX_SKILLS_SOURCE_DIR"
-	skillsTargetEnv = "GCX_SKILLS_INSTALL_DIR"
 )
 
 // embeddedSkillsFS contains the shipped skill assets.
@@ -33,13 +24,7 @@ const (
 //go:embed skills
 var embeddedSkillsFS embed.FS
 
-// skillSource describes where skill assets are loaded from.
-// In normal operation we use embedded assets; GCX_SKILLS_SOURCE_DIR is a
-// development/testing override to load from disk.
-type skillSource struct {
-	FS   fs.FS
-	Root string
-}
+var skillsManager = internalskills.NewManager(embeddedSkillsFS, "skills")
 
 // Command returns the skills command group.
 func Command() *cobra.Command {
@@ -91,14 +76,19 @@ func newListCommand() *cobra.Command {
 				return err
 			}
 
-			source, err := resolveSkillSource()
+			source, err := skillsManager.ResolveSource()
 			if err != nil {
 				return err
 			}
 
-			items, err := discoverSkills(source)
+			skills, err := skillsManager.List(source)
 			if err != nil {
 				return err
+			}
+
+			items := make([]listItem, 0, len(skills))
+			for _, s := range skills {
+				items = append(items, listItem{Name: s.Name, Description: s.Description})
 			}
 
 			return opts.IO.Encode(cmd.OutOrStdout(), items)
@@ -124,7 +114,7 @@ func (o *installOpts) setup(cmd *cobra.Command) {
 	o.IO.DefaultFormat("text")
 	o.IO.RegisterCustomCodec("text", &installTextCodec{})
 	o.IO.BindFlags(cmd.Flags())
-	cmd.Flags().StringVar(&o.Target, "dir", defaultInstallDir(), "Install directory for skills")
+	cmd.Flags().StringVar(&o.Target, "dir", skillsManager.DefaultInstallDir(), "Install directory for skills")
 	cmd.Flags().BoolVar(&o.Force, "force", false, "Overwrite the destination skill directory if it already exists")
 }
 
@@ -152,25 +142,21 @@ func newInstallCommand() *cobra.Command {
 				return err
 			}
 
-			source, err := resolveSkillSource()
+			source, err := skillsManager.ResolveSource()
 			if err != nil {
 				return err
 			}
 
 			skillName := strings.TrimSpace(args[0])
-			if skillName == "" {
-				return errors.New("skill name cannot be empty")
+			if err := internalskills.ValidateSkillName(skillName); err != nil {
+				return err
 			}
-			if strings.ContainsRune(skillName, '/') || strings.ContainsRune(skillName, '\\') {
-				return fmt.Errorf("invalid skill name %q: must not include path separators", skillName)
-			}
-
-			if !skillExists(source, skillName) {
+			if !skillsManager.Exists(source, skillName) {
 				return fmt.Errorf("skill %q not found. Run `gcx skills list` to see available skills", skillName)
 			}
 
 			targetDir := filepath.Join(opts.Target, skillName)
-			if err := installSkillDirectory(source, skillName, targetDir, opts.Force); err != nil {
+			if err := skillsManager.Install(source, skillName, targetDir, opts.Force); err != nil {
 				return err
 			}
 
@@ -183,186 +169,6 @@ func newInstallCommand() *cobra.Command {
 	}
 	opts.setup(cmd)
 	return cmd
-}
-
-func defaultInstallDir() string {
-	if fromEnv := strings.TrimSpace(os.Getenv(skillsTargetEnv)); fromEnv != "" {
-		return fromEnv
-	}
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return ".claude/skills"
-	}
-	return filepath.Join(home, ".claude", "skills")
-}
-
-func resolveSkillSource() (skillSource, error) {
-	if fromEnv := strings.TrimSpace(os.Getenv(skillsSourceEnv)); fromEnv != "" {
-		st, err := os.Stat(fromEnv)
-		if err != nil {
-			return skillSource{}, fmt.Errorf("%s is set, but path is not readable: %w", skillsSourceEnv, err)
-		}
-		if !st.IsDir() {
-			return skillSource{}, fmt.Errorf("%s must point to a directory: %s", skillsSourceEnv, fromEnv)
-		}
-		return skillSource{FS: os.DirFS(fromEnv), Root: "."}, nil
-	}
-
-	return skillSource{FS: embeddedSkillsFS, Root: "skills"}, nil
-}
-
-func discoverSkills(source skillSource) ([]listItem, error) {
-	entries, err := fs.ReadDir(source.FS, source.Root)
-	if err != nil {
-		return nil, err
-	}
-
-	items := make([]listItem, 0, len(entries))
-	for _, entry := range entries {
-		if !entry.IsDir() {
-			continue
-		}
-		skillName := entry.Name()
-		skillMD := path.Join(source.Root, skillName, "SKILL.md")
-		data, err := fs.ReadFile(source.FS, skillMD)
-		if err != nil {
-			continue
-		}
-
-		meta := readSkillMetadata(data)
-		name := skillName
-		if strings.TrimSpace(meta.Name) != "" {
-			name = strings.TrimSpace(meta.Name)
-		}
-		items = append(items, listItem{Name: name, Description: strings.TrimSpace(meta.Description)})
-	}
-
-	sort.Slice(items, func(i, j int) bool {
-		return items[i].Name < items[j].Name
-	})
-
-	return items, nil
-}
-
-func skillExists(source skillSource, skillName string) bool {
-	_, err := fs.ReadFile(source.FS, path.Join(source.Root, skillName, "SKILL.md"))
-	return err == nil
-}
-
-type skillFrontMatter struct {
-	Name        string `yaml:"name"`
-	Description string `yaml:"description"`
-}
-
-func readSkillMetadata(data []byte) skillFrontMatter {
-	content := string(data)
-	if !strings.HasPrefix(content, "---\n") {
-		return skillFrontMatter{}
-	}
-	trimmed := strings.TrimPrefix(content, "---\n")
-	frontMatter, _, ok := strings.Cut(trimmed, "\n---\n")
-	if !ok {
-		return skillFrontMatter{}
-	}
-
-	meta := skillFrontMatter{}
-	if err := yaml.Unmarshal([]byte(frontMatter), &meta); err != nil {
-		// Some skill descriptions contain informal YAML (for example unquoted ":").
-		// Fall back to a permissive parser so one bad frontmatter block does not
-		// break listing for all skills.
-		return parseFrontMatterFallback(frontMatter)
-	}
-	return meta
-}
-
-func parseFrontMatterFallback(frontMatter string) skillFrontMatter {
-	lines := strings.Split(frontMatter, "\n")
-	meta := skillFrontMatter{}
-	for i := range lines {
-		line := strings.TrimRight(lines[i], " \t")
-		trimmed := strings.TrimSpace(line)
-		switch {
-		case strings.HasPrefix(trimmed, "name:"):
-			meta.Name = strings.TrimSpace(strings.TrimPrefix(trimmed, "name:"))
-		case strings.HasPrefix(trimmed, "description:"):
-			v := strings.TrimSpace(strings.TrimPrefix(trimmed, "description:"))
-			if v == ">" || v == "|" {
-				var parts []string
-				for j := i + 1; j < len(lines); j++ {
-					next := strings.TrimRight(lines[j], " \t")
-					if strings.TrimSpace(next) == "" {
-						parts = append(parts, "")
-						continue
-					}
-					if len(next) > 0 && (next[0] == ' ' || next[0] == '\t') {
-						parts = append(parts, strings.TrimSpace(next))
-						continue
-					}
-					break
-				}
-				meta.Description = strings.TrimSpace(strings.Join(parts, " "))
-			} else {
-				meta.Description = v
-			}
-		}
-	}
-	return meta
-}
-
-// installSkillDirectory copies one embedded skill tree to a target path.
-// The copy preserves relative layout (including references/) and file modes.
-func installSkillDirectory(source skillSource, skillName, targetDir string, force bool) error {
-	if st, err := os.Stat(targetDir); err == nil && st.IsDir() {
-		if !force {
-			return fmt.Errorf("skill already installed at %s (use --force to overwrite)", targetDir)
-		}
-		if err := os.RemoveAll(targetDir); err != nil {
-			return err
-		}
-	} else if err != nil && !os.IsNotExist(err) {
-		return err
-	}
-
-	if err := os.MkdirAll(targetDir, 0o755); err != nil {
-		return err
-	}
-
-	sourceRoot := path.Join(source.Root, skillName)
-	return fs.WalkDir(source.FS, sourceRoot, func(sourcePath string, d fs.DirEntry, err error) error {
-		if err != nil {
-			return err
-		}
-		rel := strings.TrimPrefix(sourcePath, sourceRoot+"/")
-		if sourcePath == sourceRoot {
-			rel = "."
-		}
-		if rel == "." {
-			return nil
-		}
-		dstPath := filepath.Join(targetDir, filepath.FromSlash(rel))
-
-		if d.IsDir() {
-			// Use writable directory perms even for embedded assets, which are
-			// exposed as read-only and would otherwise create 0555 directories.
-			return os.MkdirAll(dstPath, 0o755)
-		}
-		if d.Type()&fs.ModeSymlink != 0 {
-			return fmt.Errorf("symlinks are not supported in skill directory: %s", sourcePath)
-		}
-
-		data, err := fs.ReadFile(source.FS, sourcePath)
-		if err != nil {
-			return err
-		}
-		info, err := d.Info()
-		if err != nil {
-			return err
-		}
-		if err := os.MkdirAll(filepath.Dir(dstPath), 0o755); err != nil {
-			return err
-		}
-		return os.WriteFile(dstPath, data, info.Mode().Perm())
-	})
 }
 
 type listTextCodec struct{}
